+++ heeyhome_webapi/heeyhomeapi/app/Http/Controllers/CollectshopController.php
@@ -47,19 +47,43 @@ class CollectshopController extends Controller
         $newpage=new PageController();
         $offset=$newpage->page($total);
         /*收藏表和店铺表两表联查*/
-        $select=DB::select('select hh_collection.id as collect_id,hh_shop.* from hh_collection 
+        $select=DB::select('select hh_shop.shop_id,hh_shop.shop_name,hh_shop.authentication,hh_shop.shop_address,hh_shop.servicearea,hh_shop.shop_scan,hh_shop.shop_volume from hh_collection 
                     left join hh_shop on hh_collection.iscollected_id=hh_shop.shop_id where collect_userid=? and collect_type=? order by collect_time desc limit ?,?',[$user_id,'shop',$offset[0],$offset[1]]);
         foreach($select as $key=>$val){
-            $img=DB::select('select shop_img from hh_shop_img where shop_id=? and is_face=?',[$val->shop_id,1]);
+            $img=DB::select('select shop_img from hh_shop_img where shop_id=? order by id desc',[$val->shop_id]);
             if($img){
                 $select[$key]->img=$img[0]->shop_img;
             }else{
                 $select[$key]->img=null;
             }
             $select[$key]->servicearea=explode(',',$val->servicearea);
-            $select[$key]->servicetag=explode(',',$val->servicetag);
-            $select[$key]->authentication=explode(',',$val->authentication);
+            $authentication=explode(',',$val->authentication);
+            foreach( $authentication as $k=>$v){
+                switch ($v){
+                    case 1: $authentication[$k]='api/public/smrz.png';
+                        break;
+                    case 2: $authentication[$k]='api/public/bzj.png';
+                        break;
+                    case 3: $authentication[$k]='api/public/tdbx.png';
+                        break;
+                    case 4: $authentication[$k]='api/public/bx.png';
+                        break;
+                    default: $authentication[$k]='';
+                        break;
+                }
+            }
+            $select[$key]->authentication=$authentication;
             $select[$key]->total=$total;
+            $score=DB::select('select projectquality,serviceattitude,overallmerit from hh_score where shop_id=?',[$val->shop_id]);
+            if($score){
+                $select[$key]->shop_score=$score[0];
+            }else{
+                $select[$key]->shop_score=array("projectquality"=>10.0,
+                    "serviceattitude"=>10.0,
+                    "overallmerit"=>10.0
+                );
+            }
+
         }
         if($select){
             $arr = array("code" => "000",
@@ -75,15 +99,16 @@ class CollectshopController extends Controller
     }
     public function del(){
         $callback=rq('callback');
+        $user_id=rq('user_id');
         $shop_id=rq('shop_id');     //店铺id
-        $del=DB::delete('delete from hh_collection where iscollected_id=?',[$shop_id]);
+        $del=DB::delete('delete from hh_collection where iscollected_id=? and  collect_userid=?',[$shop_id,$user_id]);
         if($del){
             $arr = array("code" => "000",
                 "msg" => "删除成功"
             );
             return $callback . "(" . HHJson($arr) . ")";
         }else{
-            $arr = array("code" => "117",
+            $arr = array("code" => "111",
                 "msg" => "删除失败"
             );
             return $callback . "(" . HHJson($arr) . ")";
